// project/src/data/public/programDetails.types.ts
export interface ProgramDetails {
  eyebrow: string;
  title: string;
  description: string;
  icon: string;

  duration: string;
  format: string;
  level: string;
  age: string;

  aboutTitle: string;
  aboutText: string;

  technologiesTitle: string;
  technologies: string[];

  topicsTitle: string;
  topics: string[];

  projectsTitle: string;
  projects: string[];

  resultsTitle: string;
  results: string[];

  audienceTitle: string;
  audience: string[];
}